lastest database

+++ Controller/API/NavERPController.cs
@@ -370,20 +370,41 @@ namespace FICCI_API.Controller.API
             try
             {
                 List<COAMaster> COAMaster = await GetList<COAMaster>("COA");
-                _dbContext.Database.ExecuteSqlRaw($"TRUNCATE TABLE COA_Master");
-                //foreach(var coa in COAMaster)
-                //{
-                //    CoaMaster master = new CoaMaster();
-                //    master.CoaName = 
-                //}
-                var apiResponse = new
+                if(COAMaster.Count > 0)
                 {
-                    data = COAMaster,
-                    status = true,
-                    message = $"{COAMaster.Count} records found.",
-                };
-                return Ok(apiResponse);
-
+                    var coaResponse = COAMaster.Select(c => new COAInfo
+                    {
+                        No = c.No,
+                        Name = c.Name,
+                    }).ToList();
+                    _dbContext.Database.ExecuteSqlRaw($"TRUNCATE TABLE COA_Master");
+                    foreach (var coa in coaResponse)
+                    {
+                        CoaMaster master = new CoaMaster();
+                        master.CoaNo = coa.No;
+                        master.CoaName = coa.Name;
+                        _dbContext.Add(master);
+                        _dbContext.SaveChanges();
+                    }
+                    var apiResponse = new
+                    {
+                        data = coaResponse,
+                        status = true,
+                        message = $"{COAMaster.Count} records found.",
+                    };
+                    return Ok(apiResponse);
+                }
+                else
+                {
+                    var response = new
+                    {
+                        status = true,
+                        message = "No Coa list found",
+                        data = COAMaster
+                    };
+                    return Ok(response);
+                }
+               
             }
             catch (Exception ex)
             {
@@ -397,14 +418,39 @@ namespace FICCI_API.Controller.API
             try
             {
                 List<GSTGroup> GSTGroup = await GetList<GSTGroup>("GSTGroup");
-
-                var apiResponse = new
+                if(GSTGroup.Count > 0)
                 {
-                    data = GSTGroup,
-                    status = true,
-                    message = $"{GSTGroup.Count} records found.",
-                };
-                return Ok(apiResponse);
+                    var gstResponse = GSTGroup.Select(c => new GSTGroupInfo
+                    {
+                        Code = c.Code,
+                    }).ToList();
+                    _dbContext.Database.ExecuteSqlRaw($"TRUNCATE TABLE GSTGroup");
+                    foreach (var coa in gstResponse)
+                    {
+                        Gstgroup master = new Gstgroup();
+                        master.GroupCode = coa.Code;
+                        _dbContext.Add(master);
+                        _dbContext.SaveChanges();
+                    }
+                    var apiResponse = new
+                    {
+                        data = gstResponse,
+                        status = true,
+                        message = $"{GSTGroup.Count} records found.",
+                    };
+                    return Ok(apiResponse);
+
+                }
+                else
+                {
+                    var response = new
+                    {
+                        status = true,
+                        message = "No GSTGroup list found",
+                        data = GSTGroup
+                    };
+                    return Ok(response);
+                }
 
             }
             catch (Exception ex)
@@ -419,14 +465,40 @@ namespace FICCI_API.Controller.API
             try
             {
                 List<HSNSACMaster> HSNSACMaster = await GetList<HSNSACMaster>("HSNSAC");
-
-                var apiResponse = new
+                if(HSNSACMaster.Count > 0)
                 {
-                    data = HSNSACMaster,
-                    status = true,
-                    message = $"{HSNSACMaster.Count} records found.",
-                };
-                return Ok(apiResponse);
+                    var hsnResponse = HSNSACMaster.Select(c => new HSNSACMasterInfo
+                    {
+                        Code = c.Code,
+                        GST_Group_Code = c.GST_Group_Code
+                    }).ToList();
+                    _dbContext.Database.ExecuteSqlRaw($"TRUNCATE TABLE HSNSAC");
+                    foreach(var hsn in hsnResponse)
+                    {
+                        Hsnsac master = new Hsnsac();
+                        master.HsnCode = hsn.Code;
+                        master.HsnGroup = hsn.GST_Group_Code;
+                        _dbContext.Add(master);
+                        _dbContext.SaveChanges();
+                    }
+                    var apiResponse = new
+                    {
+                        data = hsnResponse,
+                        status = true,
+                        message = $"{HSNSACMaster.Count} records found.",
+                    };
+                    return Ok(apiResponse);
+                }
+                else
+                {
+                    var response = new
+                    {
+                        status = true,
+                        message = "No HSNSAC list found",
+                        data = HSNSACMaster
+                    };
+                    return Ok(response);
+                }
 
             }
             catch (Exception ex)
@@ -441,14 +513,41 @@ namespace FICCI_API.Controller.API
             try
             {
                 List<GetLocation> GetLocation = await GetList<GetLocation>("Location");
-
-                var apiResponse = new
+                if(GetLocation.Count > 0)
                 {
-                    data = GetLocation,
-                    status = true,
-                    message = $"{GetLocation.Count} records found.",
-                };
-                return Ok(apiResponse);
+                    var loactionResponse = GetLocation.Select(c => new LoactionInfo
+                    {
+                        code = c.code,
+                        name = c.name,
+                    }).ToList();
+                    _dbContext.Database.ExecuteSqlRaw($"TRUNCATE TABLE Location");
+                    foreach (var coa in loactionResponse)
+                    {
+                        Location master = new Location();
+                        master.LocationCode = coa.code;
+                        master.LocationName = coa.name;
+                        _dbContext.Add(master);
+                        _dbContext.SaveChanges();
+                    }
+                    var apiResponse = new
+                    {
+                        data = loactionResponse,
+                        status = true,
+                        message = $"{GetLocation.Count} records found.",
+                    };
+                    return Ok(apiResponse);
+
+                }
+                else
+                {
+                    var response = new
+                    {
+                        status = true,
+                        message = "No Location list found",
+                        data = GetLocation
+                    };
+                    return Ok(response);
+                }
 
             }
             catch (Exception ex)
